# docs/agent_os_openapi.yaml
openapi: 3.1.0
info:
  title: Enstack Agent OS API
  version: 1.0.0-draft.1
  summary: Give your AI agents real money — and real control.
  description: |
    # Enstack Agent OS — /v1 API

    The Agent OS API lets a verified **Owner** (a person or business) provision
    **Agents** — AI sub-principals that hold a Balance, spend on a Visa card,
    receive into a virtual bank account, and move money over local and
    cross-border rails via **Enstack Routing** — all inside an Owner-defined
    **Policy** (limits, allowlists, human approvals, full audit).

    ## Identity model
    - **Owner** — the KYC/KYB-verified principal. Owns all funds and all liability.
    - **Agent** — a sub-principal under an Owner. No separate KYC. Every agent
      action is attributable to both the agent and its owner.
    - **Credential** — a scoped API key or MCP token bound to exactly one agent
      (or to the owner). Credentials carry scopes; scopes are further constrained
      by the agent's Policy at execution time.
    - **Policy** — the authorization layer: per-transaction / daily / monthly
      limits, counterparty and category allowlists, and human-in-the-loop
      approval rules. Policy is versioned; every evaluation is audited.

    ## Conventions
    - **Versioning**: URL prefix `/v1/...`. Breaking changes ship as `/v2`.
    - **Idempotency**: `Idempotency-Key` header is **required on every mutating
      request** (POST/PUT/PATCH). Window: 24h, keyed by `(credential, key)`.
...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /agents:
    post:
      operationId: createAgent
      tags: [Agents]
      summary: Create an agent
      description: |
        Owner scope only. Creates a sub-principal with a zero balance and the
        default (deny-all) policy. Attach a Policy before the agent can move money.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentCreate' }
      responses:
        '201':
          description: Agent created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        default: { $ref: '#/components/responses/Error' }
    get:
      operationId: listAgents
      tags: [Agents]
      summary: List agents
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/starting_after'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/AgentStatus' }
      responses:
        '200':
          description: Agent list.
          content:
            application/json:
              schema:
                type: object
                required: [data, has_more]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Agent' }
                  has_more: { type: boolean }
                  next_cursor: { type: [string, 'null'] }
        default: { $ref: '#/components/responses/Error' }
  /agents/{agent_id}:
    parameters:
      - $ref: '#/components/parameters/agentId'
    get:
      operationId: getAgent
      tags: [Agents]
      summary: Retrieve an agent
      responses:
        '200':
          description: The agent.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        default: { $ref: '#/components/responses/Error' }
    patch:
      operationId: updateAgent
      tags: [Agents]
      summary: Update agent metadata
      description: Owner scope only. Name, description, and metadata only — status changes use the suspend/reactivate actions; authorization changes live in Policy.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentUpdate' }
      responses:
        '200':
          description: Updated agent.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        default: { $ref: '#/components/responses/Error' }
  /agents/{agent_id}/suspend:
    post:
      operationId: suspendAgent
      tags: [Agents]
      summary: Suspend an agent
      description: |
        Owner scope only. Immediately freezes the agent: credentials stop
        authenticating, cards decline, in-flight payments continue to terminal
        state. Reversible via reactivate.
      parameters:
        - $ref: '#/components/parameters/agentId'
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '200':
          description: Suspended agent.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        default: { $ref: '#/components/responses/Error' }
  /agents/{agent_id}/reactivate:
    post:
      operationId: reactivateAgent
      tags: [Agents]
      summary: Reactivate a suspended agent
      parameters:
        - $ref: '#/components/parameters/agentId'
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '200':
          description: Reactivated agent.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        default: { $ref: '#/components/responses/Error' }

  # ─────────────────────────── Credentials ───────────────────────────
...
  /agents/{agent_id}/balance:
    get:
      operationId: getAgentBalance
      tags: [Balances & Transfers]
      summary: Retrieve an agent's balance
      parameters:
        - $ref: '#/components/parameters/agentId'
      responses:
        '200':
          description: The balance.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Balance' }
        default: { $ref: '#/components/responses/Error' }
  /agents/{agent_id}/funding_instructions:
    get:
      operationId: getFundingInstructions
      tags: [Balances & Transfers]
      summary: Retrieve funding instructions for an agent's balance
      description: |
        Returns the named virtual account details (account name, number,
        bank, supported rails, reference) that credit this agent's balance
        when funds arrive.
      parameters:
        - $ref: '#/components/parameters/agentId'
      responses:
        '200':
          description: Funding instructions.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FundingInstructions' }
        default: { $ref: '#/components/responses/Error' }
  /transfers:
    post:
      operationId: createTransfer
      tags: [Balances & Transfers]
      summary: Move funds between balances inside the account
      description: |
        Instant, fee-free internal ledger transfer: owner treasury ↔ agent, or
        agent ↔ agent. Policy-evaluated when initiated by an agent credential.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransferCreate' }
      responses:
        '201':
          description: Transfer completed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transfer' }
        '202':
          description: Transfer requires owner approval per Policy.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transfer' }
        default: { $ref: '#/components/responses/Error' }
    get:
      operationId: listTransfers
      tags: [Balances & Transfers]
      summary: List internal transfers
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/starting_after'
        - name: agent_id
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Transfer list.
          content:
            application/json:
              schema:
                type: object
                required: [data, has_more]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Transfer' }
                  has_more: { type: boolean }
                  next_cursor: { type: [string, 'null'] }
        default: { $ref: '#/components/responses/Error' }

  # ───────────────────────────── Quotes ─────────────────────────────
...
  /approvals:
    get:
      operationId: listApprovals
      tags: [Approvals]
      summary: List approvals
      description: Owner scope sees all; an agent credential sees only approvals raised by its own actions.
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/starting_after'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ApprovalStatus' }
        - name: agent_id
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Approval list.
          content:
            application/json:
              schema:
                type: object
                required: [data, has_more]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Approval' }
                  has_more: { type: boolean }
                  next_cursor: { type: [string, 'null'] }
        default: { $ref: '#/components/responses/Error' }
  /approvals/{approval_id}:
    get:
      operationId: getApproval
      tags: [Approvals]
      summary: Retrieve an approval
      parameters:
        - name: approval_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The approval.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        default: { $ref: '#/components/responses/Error' }
  /approvals/{approval_id}/approve:
    post:
      operationId: approveApproval
      tags: [Approvals]
      summary: Approve a pending action
      description: Owner scope only. Releases the held action (payment, transfer, …) for execution at the current quote/route — if the underlying quote has expired, the action is repriced and the approval records both amounts.
      parameters:
        - name: approval_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string, maxLength: 500 }
      responses:
        '200':
          description: Approved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        default: { $ref: '#/components/responses/Error' }
  /approvals/{approval_id}/reject:
    post:
      operationId: rejectApproval
      tags: [Approvals]
      summary: Reject a pending action
      description: Owner scope only. The held action transitions to `cancelled` with reason `owner_rejected`.
      parameters:
        - name: approval_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string, maxLength: 500 }
      responses:
        '200':
          description: Rejected.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        default: { $ref: '#/components/responses/Error' }

  # ──────────────────────────── Activity ────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Owner key (`ok_live_…`) or agent credential (`ac_live_…`).
        Agent credentials are policy-bound: server-side Policy evaluation
        applies to every call regardless of credential scopes.
  parameters:
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated unique key (UUIDv4 recommended). Required on every
        mutating request. 24h dedup window keyed by (credential, key).
        Same key + same body replays the original response; same key +
        different body returns 409.
      schema:
        type: string
        minLength: 16
        maxLength: 128
    agentId:
      name: agent_id
      in: path
      required: true
      schema: { type: string }
      description: Agent identifier, e.g. `agt_01J….`
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    starting_after:
      name: starting_after
      in: query
      schema: { type: string }
      description: Cursor — the `next_cursor` value from the previous page.
  responses:
    Error:
      description: |
        Error envelope. Notable codes: `policy_denied` (403, hard policy
        block), `insufficient_funds` (409), `idempotency_conflict` (409,
        key reuse with different body), `quote_expired` (409),
        `approval_required` is NOT an error — held actions return 202.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: Machine-readable error code.
          examples: [policy_denied, insufficient_funds, quote_expired, idempotency_conflict, not_found, unauthorized, validation_error, rate_limited]
        message:
          type: string
          description: Human-readable description.
        details:
          type: [object, 'null']
          description: Field-level or context detail (e.g. which policy rule denied).
          additionalProperties: true
        request_id:
          type: string
          description: Correlate with support and the activity feed.
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^-?\d+(\.\d+)?$'
          description: Decimal string. Never a float.
          examples: ["1500.00", "0.4275"]
        currency:
          type: string
          description: ISO 4217 code.
          examples: [USD, PHP, HKD, EUR, SGD]
...
    AgentStatus:
      type: string
      enum: [active, suspended, deactivated]
    Agent:
      type: object
      required: [id, owner_id, display_name, status, policy_version, created_at]
      properties:
        id: { type: string, examples: [agt_01JABCDXYZ] }
        owner_id: { type: string }
        display_name:
          type: string
          description: Appears on virtual account naming and card descriptors.
        description: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/AgentStatus' }
        policy_version:
          type: integer
          description: Version number of the active policy (0 = default deny-all).
        metadata:
          type: object
          additionalProperties: { type: string }
          description: Free-form key-value store (max 40 keys).
        created_at: { type: string, format: date-time }
    AgentCreate:
      type: object
      required: [display_name]
      properties:
        display_name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500 }
        metadata:
          type: object
          additionalProperties: { type: string }
    AgentUpdate:
      type: object
      properties:
        display_name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500 }
        metadata:
          type: object
          additionalProperties: { type: string }
...
    Balance:
      type: object
      required: [holder_type, holder_id, available, currency_breakdown, updated_at]
      properties:
        holder_type: { type: string, enum: [owner, agent] }
        holder_id: { type: string }
        available:
          $ref: '#/components/schemas/Money'
          description: Total available, expressed in the owner default currency.
        pending_out:
          $ref: '#/components/schemas/Money'
          description: Reserved by in-flight payments and pending approvals.
        currency_breakdown:
          type: array
          items: { $ref: '#/components/schemas/Money' }
          description: Per-currency components of the balance.
        updated_at: { type: string, format: date-time }
    FundingInstructions:
      type: object
      required: [agent_id, methods]
      properties:
        agent_id: { type: string }
        methods:
          type: array
          items:
            type: object
            required: [rail, account_name, reference]
            properties:
              rail:
                type: string
                description: Funding rail identifier.
                examples: [bank_transfer_local, bank_transfer_international, fps, sepa_instant, ach]
              account_name: { type: string }
              account_number: { type: [string, 'null'] }
              bank_name: { type: [string, 'null'] }
              bank_identifier:
                type: [string, 'null']
                description: SWIFT/BIC, sort code, or local bank code as applicable.
              currency: { type: string }
              reference:
                type: string
                description: Must be included with the transfer for automatic crediting.
    Transfer:
      type: object
      required: [id, source, destination, amount, status, created_at]
      properties:
        id: { type: string, examples: [trf_01JABCDXYZ] }
        source: { $ref: '#/components/schemas/BalanceRef' }
        destination: { $ref: '#/components/schemas/BalanceRef' }
        amount: { $ref: '#/components/schemas/Money' }
        status: { type: string, enum: [completed, pending_approval, cancelled] }
        approval_id: { type: [string, 'null'] }
        note: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
    BalanceRef:
      type: object
      required: [holder_type, holder_id]
      properties:
        holder_type: { type: string, enum: [owner, agent] }
        holder_id: { type: string }
    TransferCreate:
      type: object
      required: [source, destination, amount]
      properties:
        source: { $ref: '#/components/schemas/BalanceRef' }
        destination: { $ref: '#/components/schemas/BalanceRef' }
        amount: { $ref: '#/components/schemas/Money' }
        note: { type: string, maxLength: 280 }
...
    ApprovalStatus:
      type: string
      enum: [pending, approved, rejected, expired]
    Approval:
      type: object
      required: [id, agent_id, action_type, action_id, status, created_at]
      properties:
        id: { type: string, examples: [apr_01JABCDXYZ] }
        agent_id: { type: string }
        action_type: { type: string, enum: [payment, transfer, counterparty] }
        action_id: { type: string }
        trigger:
          type: string
          description: Which policy rule raised the approval.
          examples: [require_approval_above, new_counterparty, cross_border]
        summary:
          type: object
          description: Denormalized snapshot of the held action, rendered on the owner's approval card.
          additionalProperties: true
        status: { $ref: '#/components/schemas/ApprovalStatus' }
        decided_by: { type: [string, 'null'] }
        decided_at: { type: [string, 'null'], format: date-time }
        note: { type: [string, 'null'] }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
...